IS: Move files to correct folders
Så et behov for å rydde litt i ispersonoppgave. Her er det både litt flytting av generator-filer og navnendringer.
