+++ src/main/kotlin/no/nav/syfo/aktivitetskrav/kafka/domain/ExpiredVarsel.kt
@@ -1,4 +1,4 @@
-package no.nav.syfo.aktivitetskrav.domain
+package no.nav.syfo.aktivitetskrav.kafka.domain
 
 import no.nav.syfo.domain.PersonIdent
 import java.time.LocalDate
