+++ src/test/kotlin/no/nav/syfo/dialogmotestatusendring/DialogmoteStatusendringServiceSpek.kt
@@ -3,7 +3,9 @@ package no.nav.syfo.dialogmotestatusendring
 import io.mockk.*
 import no.nav.syfo.dialogmotestatusendring.domain.DialogmoteStatusendringType
 import no.nav.syfo.personoppgave.*
-import no.nav.syfo.testutil.*
+import no.nav.syfo.testutil.generators.generateDialogmotestatusendring
+import no.nav.syfo.testutil.generators.generatePPersonoppgave
+import no.nav.syfo.testutil.generators.generatePersonoppgave
 import no.nav.syfo.util.toLocalDateTimeOslo
 import org.spekframework.spek2.Spek
 import org.spekframework.spek2.style.specification.describe
