+++ src/test/kotlin/no/nav/syfo/testutil/generators/KafkaIdenthendelseDTOGenerator.kt
@@ -1,9 +1,10 @@
-package no.nav.syfo.testutil
+package no.nav.syfo.testutil.generators
 
 import no.nav.syfo.domain.PersonIdent
 import no.nav.syfo.identhendelse.kafka.IdentType
 import no.nav.syfo.identhendelse.kafka.Identifikator
 import no.nav.syfo.identhendelse.kafka.KafkaIdenthendelseDTO
+import no.nav.syfo.testutil.UserConstants
 
 fun generateKafkaIdenthendelseDTO(
     personident: PersonIdent = UserConstants.ARBEIDSTAKER_FNR,
