+++ src/test/kotlin/no/nav/syfo/testutil/generators/KafkaMeldingDTOGenerator.kt
@@ -1,8 +1,9 @@
-package no.nav.syfo.testutil
+package no.nav.syfo.testutil.generators
 
 import no.nav.syfo.domain.PersonIdent
 import no.nav.syfo.behandlerdialog.domain.KMeldingDTO
 import no.nav.syfo.behandlerdialog.domain.MeldingType
+import no.nav.syfo.testutil.UserConstants
 import java.time.OffsetDateTime
 import java.util.*
